Added benchmark for 2020-01
