+++ 2020-01/opcode0x90/Cargo.toml
@@ -6,5 +6,22 @@ edition = "2018"
 
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
+[lib]
+name = "aoc"
+path = "src/lib.rs"
+bench = false
+
+[[bin]]
+name = "aoc"
+path = "src/main.rs"
+bench = false
+
+[[bench]]
+name = "bench"
+harness = false
+
 [dependencies]
 itertools = "0.9"
+
+[dev-dependencies]
+criterion = "0.3"
